refactored getTeams path
