+++ models/db/dbGetTeams.js
@@ -1,0 +1,6 @@
+import query from "../../dbConfig/index.js"
+
+export async function dbGetTeams(){
+    const teams = await query(`SELECT * FROM teams;`)
+    return teams.rows
+}
